Flake8 code cleanup
