+++ insteonplm/devices/sensorsActuators.py
@@ -1,14 +1,17 @@
-from .devicebase import DeviceBase
-from insteonplm.constants import *
+"""INSTEON Sensors Actuators Device Class Module."""
+from insteonplm.devices import DeviceBase
 from insteonplm.states.onOff import OpenClosedRelay
 from insteonplm.states.sensor import IoLincSensor
+
 
 class SensorsActuators(DeviceBase):
     """Sensors And Actuator Device Class 0x07
 
      There are 3 known device types in this category:
      1) I/O Linc [2450] & [2450-50-60]
-     2) Smartenit IO devices (various input and output channels, see http://sandbox.smartenit.com/downloads/IoTxx_Command_Set.pdf) Including
+     2) Smartenit IO devices (various input and output channels,
+     see http://sandbox.smartenit.com/downloads/IoTxx_Command_Set.pdf)
+     Including
         a) EZSns1W Sensor Interface Module
         b) EZIO8T I/O Module
         c) EZIO2X4
@@ -17,23 +20,30 @@ class SensorsActuators(DeviceBase):
         f) EZISnsRf
         g) EZIO6I
         h) EZIO4O
-    3) SynchroLinc [2423A5] (http://cache.insteon.com/developer/2423A5dev-112010-en.pdf) 
+    3) SynchroLinc [2423A5]
+    (http://cache.insteon.com/developer/2423A5dev-112010-en.pdf)
 
-    Each device type is sufficiently different as to require their own device class.
-    However, they all seem to have a common element of a relay and a sensor. 
+    Each device type is sufficiently different as to require their own device
+    class. However, they all seem to have a common element of a relay and a
+    sensor.
     """
-    def __init__(self, plm, address, cat, subcat, product_key = 0, description = '', model = ''):
-        return super().__init__(plm, address, cat, subcat, product_key, description, model)
+    def __init__(self, plm, address, cat, subcat, product_key=0,
+                 description='', model=''):
+        super().__init__(plm, address, cat, subcat, product_key,
+                         description, model)
 
-        self._stateList[0x01] = OpenClosedRelay(self._address, "openClosedRelay", 0x01, self._send_msg, self._plm.message_callbacks, 0x00)
+        self._stateList[0x01] = OpenClosedRelay(
+            self._address, "openClosedRelay", 0x01, self._send_msg,
+            self._plm.message_callbacks, 0x00)
+
 
 class SensorsActuators_2450(SensorsActuators):
     """I/O Linc [2450] & [2450-50-60] Device Class 0x07 subcat 0x00
-        
+
     Two separate INSTEON devices are created
         1) Relay
             - ID: xxxxxx (where xxxxxx is the Insteon address of the device)
-            - Controls: 
+            - Controls:
                 - relay_close()
                 - relay_open()
             - Monitor: relay.connect(callback)
@@ -50,8 +60,14 @@ class SensorsActuators_2450(SensorsActuators):
         - callback(self, device_id, state, state_value)
     """
 
-    def __init__(self, plm, address, cat, subcat, product_key=None, description=None, model=None):
-        super().__init__(plm, address, cat, subcat, product_key, description, model)
+    def __init__(self, plm, address, cat, subcat, product_key=None,
+                 description=None, model=None):
+        super().__init__(plm, address, cat, subcat, product_key,
+                         description, model)
 
-        self._stateList[0x01] = OpenClosedRelay(self._address, "openClosedRelay", 0x01, self._send_msg, self._plm.message_callbacks, 0x00)
-        self._stateList[0x02] = IoLincSensor(self._address, "openClosedSensor", 0x02, self._send_msg, self._plm.message_callbacks, 0x00)
+        self._stateList[0x01] = OpenClosedRelay(
+            self._address, "openClosedRelay", 0x01, self._send_msg,
+            self._plm.message_callbacks, 0x00)
+        self._stateList[0x02] = IoLincSensor(
+            self._address, "openClosedSensor", 0x02, self._send_msg,
+            self._plm.message_callbacks, 0x00)
